Code_Auditor/requirements.txt
langchain==1.2.0
langchain-openai==1.1.6
langchain-community>=0.3.0
langchain-classic
pydantic>=2.7.0
nbformat==5.10.4
ipywidgets>=8.1.0
rich>=13.7.0
ipykernel
python-dotenv
nbformat
